+++ tests/default.nix
@@ -40,6 +40,7 @@ let
     // prefix "host" (importTests "host.nix")
     // prefix "vlanId" (importTests "vlan-id.nix")
     // prefix "mtu" (importTests "mtu.nix")
+    // prefix "icmpType" (importTests "icmp-type.nix")
     // prefix "registry" (importTests "registry.nix")
     // prefix "iparse" (importTests "internal/parse.nix")
     // prefix "ifmt" (importTests "internal/format.nix")